Move warehouse.macaroons.caveats into a package
